Update Justtheaudiostuff3_sounddevice.py
Grain Size in Milliseconds:

Grain size will be specified in milliseconds and converted to samples based on the sample rate (fs). We will allow users to adjust the grain size in increments of 50 ms, with a minimum grain size of 50 ms.
Softened Default Envelope:

We'll create a softer envelope by default and provide the user with the option to switch between different envelope types using a keypress.
Key Assignments for Grain Size, Envelope, and Random Variation:

We'll provide keys to increase and decrease grain size, change the envelope type, and adjust random variation in grain size.
Random Variation Around Grain Size:

Introduce a random variation factor (expressed as a percentage) that can add randomness to the grain size. Users can adjust this factor in 50% increments using assigned keys.
Key Mapping Display:

Print out all the key assignments at the start of the program.